+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.6.0
 - Use default OS config directory instead of home directory to read the configuration file by default.
+- Add `init` subcommand that creates necessary directories and adds a configuration file.
 
 # 0.5.0
 - If at least one task fails to build **pkger** will now return with exit code 1 [#66](https://github.com/vv9k/pkger/pull/66)